Feat #13
current_orders has been eliminated. A delivery man doesnt need to know all of his orders. Since the manager is
going to reasign orders to different delivery men depending on the
situation seeying the orders change constantly could confuse a delivery
man. The only information he is going to be shown is what his next
destination is and what he has to do there.

All atributes need to be private.

Weigth can optionally be assigned in the constructor since there can be
an standard and also exceptions.

updateRoute(...) needs to notify the delivery man if the new next
destination is different from the previous one.