Feat #13
Since there is not going to be a map in the class because in this
problem there is no access to a real map in which routes can be
calculated and new client locations are constantly appearing and it
isn't worth it to store the old clients location since they are not
going to configure a map in any case a more appropeate name for the
class is the given.

The route_calculator needs to know all delivery men at its disposal and
it knows all the orders since an order is assign to a delivery man at
any given time.