Feat #20
A route has associated a number of orders and a set of destinations.
There is no need to exlain why it has associated a set of destinations.
It has associated a set of orders because, that way, the process of
updating the route can be automated; when a deliveryMan gets to the
location of a restaurant he will be able to call a method to receive the
products associated with all the current orders related to that
restaurant. After that, he will succesfully receive the products. There
will be similas function to deliver orders. That way, because the route
knows the orders related to such deliveryMan it will check if there are
any products in the orders wich origin the current next location or any
orders with destination the current next location. If there are any that
means the deliveryMan still has something to do in that location,
otherwise it means that all that was to be done in such location is
finished and he is ready to head to the next location, hence the route
gets updated.