Feat #20
There is a default maximumWeight since many deliveryMan use the same
transport method(bike), hence will be able to load the same ammount of
weight.

It is absolutly necessary to know the location of a deliveryMan at any
given moment since it will be used to calculate the routes. Since there
is going to be no map coded in the problem because it doesnt make sense
it becomes necessary to code such information if the way of an
attribute: coordinates.

There is no order list related to a deliveryMan directly because it is
related to the routes. That way, since there is only one route related
to a deliveryMan and one list of products related to a route there is a
univocal conexion between the deliveryMan and its orders.