Feat #20
State cannot be related to the order, it has to be related to the
products of the order. Since not all products have to be from the same
restaurant, there is no way to know if an order is ready in such cases
cause there will be a momment in which some products will be ready and
already picked up and some othes won't and this situation can't be
represented if the state is related to the order.