Objetivo 4 (#19)
* Changed ts-matrix

* Eliminate boilerplate in tsconfig.json

* Fix tsconfig

* .

* Set of changes:
Change variables names to english
Change variables scopes
Implemented file reading

* Added mocha and chai to the developer dependencies

* Feat #18

* Feat #18

The file format was decided to be a json because it makes reading the
file easer.
tryReadFile: An exception is raised if no file is found at the parameter route
tryParseJson: An exception is raised if unable to parse the parameter
string

* Add Tests for read_files

* Feat #18

Little nomenclature fix

* Feat #18

loadLocations() directly replaced with loadRestaurants() because there
aren't any other locations to load

* Feat #18

* .

* .

* .

* .

* .

* Feat #13

current_orders has been eliminated. A delivery man doesnt need to know all of his orders. Since the manager is
going to reasign orders to different delivery men depending on the
situation seeying the orders change constantly could confuse a delivery
man. The only information he is going to be shown is what his next
destination is and what he has to do there.

All atributes need to be private.

Weigth can optionally be assigned in the constructor since there can be
an standard and also exceptions.

updateRoute(...) needs to notify the delivery man if the new next
destination is different from the previous one.

* Feat #18

Orders need to have an state since it can be used to inform clients of
the current state of the product.

* Feat #13, #18

Everything related to the routes is done in the route_calculator, thus
there is no need for manager to contain list of routes.

If deleting an inexistant restaurant is attempted an exception is
raised.

When a new order is recieved the assigned delivery man's route needs to
be updated.

* Feat #13

Since there is not going to be a map in the class because in this
problem there is no access to a real map in which routes can be
calculated and new client locations are constantly appearing and it
isn't worth it to store the old clients location since they are not
going to configure a map in any case a more appropeate name for the
class is the given.

The route_calculator needs to know all delivery men at its disposal and
it knows all the orders since an order is assign to a delivery man at
any given time.

* Feat #13

Since there is not going to be a map in the class because in this
problem there is no access to a real map in which routes can be
calculated and new client locations are constantly appearing and it
isn't worth it to store the old clients location since they are not
going to configure a map in any case a more appropeate name for the
class is the given.

The route_calculator needs to know all delivery men at its disposal and
it knows all the orders since an order is assign to a delivery man at
any given time.

addOrder(...) exists since there might be a necessity to do some
operations before calling recalculateRoutes().

* Feat #13

An order has to be assigned to one and only one delivery_man at any
given moment, thus it makes sense to have them stored in their
correspondant delivery_man.

Everytime a route is changed the delivery_man needs to be updated.

* Feat #13

RouteCalculator is in charge to modify a change route of a delivery man.

* Feat #13

Since there is on map is important to know current delivery_man
location.

* Feat #13

A route is going to have all delivery man orders assign since it is a
one to one relationship and it has many benefits since calling update it
can be known if all orders related to the next location are done(since
orders have states).

* Feat #13

IDs start at 1.

* .

* Feat #13

Array lenght can be forced in typescript

* Feat #13

.

* Build directory should have never been in repository it is generated.

* Mocha doesnt generate .js files

* Feat #13

Fix typo

* Test descriptions must be predicates no gerunds

* .

* .

* .

* .

* Feat #18

Data read has to be asyncronous

* Tests neede to be updated since only some of them are asyncronous

* Packages are needed for async tests

* Unused

* Feat #20

A route has associated a number of orders and a set of destinations.
There is no need to exlain why it has associated a set of destinations.
It has associated a set of orders because, that way, the process of
updating the route can be automated; when a deliveryMan gets to the
location of a restaurant he will be able to call a method to receive the
products associated with all the current orders related to that
restaurant. After that, he will succesfully receive the products. There
will be similas function to deliver orders. That way, because the route
knows the orders related to such deliveryMan it will check if there are
any products in the orders wich origin the current next location or any
orders with destination the current next location. If there are any that
means the deliveryMan still has something to do in that location,
otherwise it means that all that was to be done in such location is
finished and he is ready to head to the next location, hence the route
gets updated.

* Feat #20

There is a default maximumWeight since many deliveryMan use the same
transport method(bike), hence will be able to load the same ammount of
weight.

It is absolutly necessary to know the location of a deliveryMan at any
given moment since it will be used to calculate the routes. Since there
is going to be no map coded in the problem because it doesnt make sense
it becomes necessary to code such information if the way of an
attribute: coordinates.

There is no order list related to a deliveryMan directly because it is
related to the routes. That way, since there is only one route related
to a deliveryMan and one list of products related to a route there is a
univocal conexion between the deliveryMan and its orders.

* Feat #20

A state is needed to be related to a product in order for the routes to
automatically update.

* Feat #20

State cannot be related to the order, it has to be related to the
products of the order. Since not all products have to be from the same
restaurant, there is no way to know if an order is ready in such cases
cause there will be a momment in which some products will be ready and
already picked up and some othes won't and this situation can't be
represented if the state is related to the order.

* Feat #20

routeCalculator only needs to know the current delivery men; Since an
order is alwais going to be related to a route wich is related to a
delivery man, routes can be accesed at any given moment and that way
current orders and their state too.

* Fix typo

* .

* .

* Fix typo

* .

* Feat #20

Se ha implementado la distancia euclidea al cuadrado porque igualmente
es una distancia pero computacionalmente es mas mucho mas liviana.

* Fix typo

* .

* Feat #20

Restaurant needs to be in its own file since it is a different class
from Location.

* Feat #20

Tests are now in the root folder instead of src since they really arent
part of the source code.

* Fix compilation config

* Feat #20

I/O has to be asyncronous.

* Feat #20

An order can be compromised by products from many different different
restaurants, thus origin must be an array.

The only interesting data that has to be stored in origin is the
location and id of the restaurants not the restaurant objects
themselves.

* Feat #20

Better identifier since there is no change in the state of the object.